requirements.txt
absl-py==1.4.0
addict==2.4.0
aiosignal==1.3.1
albumentations==1.0.3
alembic==1.9.4
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
autoattack @ git+https://github.com/fra31/auto-attack.git@9b264b52bb65c373373727f532865a5551ab9c02
autopage==0.5.1
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
brotlipy==0.7.0
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
chardet==3.0.4
charset-normalizer==3.0.1
click==8.0.4
cliff==4.2.0
cmaes==0.9.1
cmd2==2.4.3
colorama==0.4.6
colorlog==6.7.0
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
easydict==1.10
elasticdeform==0.4.9
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
flatbuffers==23.1.21
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
frozenlist==1.3.3
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
GPUtil==1.4.0
greenlet==2.0.2
grpcio==1.51.3
h5py==3.8.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
iniconfig==1.1.1
intel-cmplr-lib-rt==2023.0.0
intel-cmplr-lic-rt==2023.0.0
intel-openmp==2023.0.0
iopath==0.1.10
ipykernel==6.21.2
ipython==8.11.0
ipython-genutils==0.2.0
jedi==0.18.2
Jinja2==2.11.3
joblib==1.2.0
json-tricks==3.15.5
json5==0.9.10
jsonschema==4.17.3
jupyter-server==1.23.1
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.2
keras==2.11.0
kiwisolver==1.4.4
lazy_loader==0.1
libclang==15.0.6.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mpmath==1.2.1
msgpack==1.0.4
munkres==1.1.4
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.4
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
nvidia-ml-py==11.495.46
oauthlib==3.2.2
opencv-python==4.6.0.66
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
optuna==3.0.2
ordered-set==4.1.0
packaging==23.0
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
patsy==0.5.3
pbr==5.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.0.0
pluggy==1.0.0
portalocker==2.7.0
prettytable==3.4.1
progress==1.6
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyflakes==2.5.0
pyglet==1.5.27
Pygments==2.14.0
PyOpenGL==3.1.0
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.18.1
PySocks==1.7.1
pytest-runner==6.0.0
python-dateutil==2.8.2
python-slugify==6.1.2
-e git+https://github.com/dali-dl/tada@91a8be11513992106b2e84783c9f65fcfe3cbe5e#egg=pytorch_adapt
pytorch-metric-learning==1.6.2
pytz==2022.4
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
ray==2.0.0
requests==2.28.2
requests-oauthlib==1.3.1
robustbench @ git+https://github.com/robustbench/robustbench@4d73272b1c6017b98b33148eaea93e5d71c311e0
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
Send2Trash==1.8.0
setproctitle==1.3.2
Shapely==1.8.4
shortuuid==1.0.9
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.4
stack-data==0.6.2
statsmodels==0.13.5
stevedore==5.0.0
sympy==1.11.1
tabulate==0.8.10
tbb==2021.8.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.1
terminaltables==3.1.10
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2023.2.28
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchmetrics==0.10.0
torchsummary==1.5.1
torchvision==0.13.1
tornado==6.2
tqdm==4.56.2
traitlets==5.9.0
trimesh==3.15.3
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
urllib3==1.26.11
virtualenv==20.20.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.4.1
-e git+https://github.com/matthijsz/weightedcorr@680ccdefa50d363ab14988080d8ac89ba6aad3ab#egg=WeightedCorr
Werkzeug==2.2.3
wrapt==1.15.0
xdoctest==1.1.0
xgboost==1.7.4
xmltodict==0.13.0
yacs==0.1.8
yapf==0.32.0
zipp==3.8.1
